Add new sensor definitions and processing functions for Ducobox components
- Introduced BOX_SENSORS, COMMBOARD_SENSORS, DUCONETWORK_SENSORS, NODE_SENSORS, and CALIBRATION_SENSORS to enhance sensor capabilities.
- Implemented processing functions for temperature, humidity, CO₂, pressure, and network status.
- Refactored sensor entity descriptions for improved clarity and consistency.
