+++ custom_components/duco_ventilation_sun_control/common.py
@@ -1,0 +1,88 @@
+def _process_node_temperature(value):
+    """Process node temperature values."""
+    if value is not None:
+        return value  # Assuming value is in Celsius
+    return None
+
+def _process_node_humidity(value):
+    """Process node humidity values."""
+    if value is not None:
+        return value  # Assuming value is in percentage
+    return None
+
+def _process_node_co2(value):
+    """Process node CO₂ values."""
+    if value is not None:
+        return value  # Assuming value is in ppm
+    return None
+
+def _process_node_iaq(value):
+    """Process node IAQ values."""
+    if value is not None:
+        return value  # Assuming value is in percentage
+    return None
+
+# Main sensor processing functions
+def _process_temperature(value):
+    """Process temperature values by dividing by 10."""
+    if value is not None:
+        return value / 10.0  # Convert from tenths of degrees Celsius
+    return None
+
+def _process_speed(value):
+    """Process speed values."""
+    if value is not None:
+        return value  # Assuming value is already in RPM
+    return None
+
+def _process_pressure(value):
+    """Process pressure values."""
+    if value is not None:
+        return value  # Assuming value is in Pa
+    return None
+
+def _process_rssi(value):
+    """Process Wi-Fi signal strength."""
+    if value is not None:
+        return value  # Assuming value is in dBm
+    return None
+
+def _process_uptime(value):
+    """Process device uptime."""
+    if value is not None:
+        return value  # Assuming value is in seconds
+    return None
+
+def _process_timefilterremain(value):
+    """Process filter time remaining."""
+    if value is not None:
+        return value  # Assuming value is in days
+    return None
+
+def _process_bypass_position(value):
+    """Process bypass position."""
+    if value is not None:
+        # Assuming value ranges from 0 to 255, where 255 is 100%
+        return round((value / 255) * 100)
+    return None
+
+def _process_network_status(value):
+    """Process network status."""
+    if value is not None:
+        return value.title()  # Assuming value is a string
+    return None
+
+def _process_calibration_status(value):
+    """Process calibration status."""
+    if value is not None:
+        if value:
+            return "Valid"
+        else:
+            return "Invalid"
+    return None
+
+def _process_calibration_state(value):
+    """Process calibration state."""
+    if value is not None:
+        return value.title()  # Assuming value is a string
+    return None
